% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyEquilibrium.R
\name{classify_equilibrium}
\alias{classify_equilibrium}
\title{Classify equilibrium points of a dynamical system}
\usage{
classify_equilibrium(A)
}
\arguments{
\item{A}{the jacobian matrix evaluated at a given equilibrium point (model dependent)}
}
\value{
list
}
\description{
{Given the Jacobian matrix evaluated at a fixed point, this function classifies the stability and whether it is a sink or source.}
}
\details{
{The classification scheme is based off the Poincare diagram in the \eqn{Tr(A), det(A)} plane. The user must know the form of the Jacobian matrix
as well as the fixed points and be able to compute the Jacobian at each of the fixed points. These are heavily model dependent.}
}
