% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odeEulerScheme.R
\name{ode.EulerScheme}
\alias{ode.EulerScheme}
\title{Solve arbitrary first order ODE systems}
\usage{
ode.EulerScheme(dynamics, IC, t0 = 0, tn = 1, n = 1000)
}
\arguments{
\item{dynamics}{a list of functions with as many variables as IC, and in the same order with the names defining the dynamics of the system}

\item{IC}{a list of variables with the same names as the non-time input as dynamics, in the same order with initial values}

\item{t0}{initial time}

\item{tn}{ending time}

\item{n}{number of sub-intervals in time-grid}
}
\value{
data.frame
}
\description{
{A general Euler scheme implementation for arbitrary first-order ODE systems of finite dimension.}
}
\details{
{The list of functions each element with arguments \code{(t,x,y,z)} for each element matching the elements of \code{IC} as \code{list(x = y0, y = y0, z = z0)} for example.}
}
