% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{ode}
\alias{ode}
\title{Solve arbitrary first order ODE systems}
\usage{
ode(f, IC, parameters, t0 = 0, tn = 1, n = 1000, engine = "RK4")
}
\arguments{
\item{f}{a list of functions with as many variables as IC, and in the same order with the names}

\item{IC}{a list of variables with the same names as the non-time input as f, in the same order with initial values}

\item{parameters}{a list of auxillary parameters of the model defined globally (it is not actually used in the function call...)}

\item{t0}{initial time}

\item{tn}{ending time}

\item{n}{number of sub-intervals in time-grid}

\item{engine}{which scheme to use to solve the ODE; 'RK4' or 'Euler', see details.}
}
\value{
data.frame
}
\description{
{A solver for arbitrary first-order ODE systems of finite dimension, using either RK4 (default) or an Euler scheme.}
}
\details{
{The list of functions must have syntax e.g. \code{f(t,x,y,z)} for each element matching the named elements of \code{IC} as \code{list(x = y0, y = y0, z = z0)} for example.
Further, the engine must be either "RK4" for a fourth-order Runge-Kutta method, or "Euler" for an Euler scheme.}
}
