% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode1.R
\name{ode1}
\alias{ode1}
\title{Solve for first order ODEs}
\usage{
ode1(f, y0, t0 = 0, tn = 1, n = 1000, engine = "RK4")
}
\arguments{
\item{f}{the function of t and y in the explicit ODE \eqn{y'=f(t,y)}}

\item{y0}{the initial value}

\item{t0}{the initial value input}

\item{tn}{the terminal value input}

\item{n}{the number of sub-intervals to use, defaults to 1000}

\item{engine}{the type of scheme to use, defaults to RK4.}
}
\value{
data.frame of the time grid and solution grid
}
\description{
{One dimensional first order ODE solver using either RK4 or an Euler scheme..}
}
\details{
{\code{f} must be a function of time and the state variable \code{y}. The engine must either be "RK4" or "Euler"}
}
