% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectories.R
\name{plot_trajectories}
\alias{plot_trajectories}
\title{Plot state-variables trajectory over time given a numerical solution}
\usage{
plot_trajectories(
  sol,
  legend_names = NULL,
  legend_loc = "topright",
  legend_size = 0.9
)
}
\arguments{
\item{sol}{solution to plot, object returned from \code{ode}, \code{ode1} etc}

\item{legend_names}{optional, will otherwise use the names given to \code{sol}}

\item{legend_loc}{where to place the legend: "topleft", "topright", etc.}

\item{legend_size}{number between 0 and 1 for scaling the legend box}
}
\value{
null
}
\description{
{Function for plotting trajectories of the solution to an ODE system.}
}
