% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidarthe.R
\name{sidarthe}
\alias{sidarthe}
\title{Euler scheme for the SIDARTHE model}
\usage{
sidarthe(
  parameters,
  initial_conditions,
  t0 = 0,
  tn = 1,
  n = 1000,
  verbose = TRUE
)
}
\arguments{
\item{parameters}{named list of parameters see details}

\item{initial_conditions}{named list of initial conditions, see details}

\item{t0}{initial time point}

\item{tn}{the terminal time point}

\item{n}{number of sub-intervals in time-grid}

\item{verbose}{whether to plot the populations over time and print parameter+reproductive ratio}
}
\value{
list consisting of (1) parameters, a data.frame containing the contact rates, etc and (2) sidarthe, a matrix whose columns consist of the
eight (fractions of) populations over time:
\itemize{
\item suspected
\item infected
\item diagnosed
\item ailing
\item recognized
\item threatened
\item healed
\item extinct}
and finally (3) the basic reproductive ratio as defined for this extended SIR model as a consequence of Proposition 2 in the original paper.
}
\description{
{A basic Euler scheme is used to numerically solve the ODE system of the novel SIDARTHE model.}
}
\details{
{The parameters must be a named list containing values
\itemize{
\item \eqn{\alpha}, \eqn{\beta}, \eqn{\gamma}, \eqn{\delta} "represent the transimission rate
(i.e.  the probability of disease transmission in a singlecontact times the average number of contacts
per person) due to contacts between a Susceptible subject and an Infected, a Diagnosed, an Ailing,
a Recognised subject. Typically, \eqn{\alpha} is larger than \eqn{\gamma} (assuming that people tend to
avoid contacts with subjects showing symptoms, even though diagnosis has not been made yet), which in turn
is probably larger than \eqn{\beta} and \eqn{\delta} (assuming that subjects who have been diagnosed are
properly isolated). These parameters can be modified by social distancing policies (e.g., closing schools,
remote working, etc.). The risk of contagion due to Threatened subjects, treated in proper ICUs, is assumed negligible."

\item \eqn{\epsilon} and \eqn{\theta} "capture capture  the  probability  rate  of  detection,  relative
to  asymptomatic  and  mildly  symptomatic  casesrespectively.  These parameters, also modifiable,
reflect the level of attention on the disease and the numberof tests performed over the population. Note that
\eqn{\theta} is typically larger than \eqn{\epsilon}, since symptomatic individuals are more likely to get tested."

\item \eqn{\zeta} and \eqn{\eta} "denote the probability rate at which an infected subject,
respectively not aware and aware of being infected, develops clinically relevant symptoms, and
are probably comparable.  These parameters are disease-dependent and hardly modifiable."

\item \eqn{\mu} and \eqn{\nu} "respectively denote the rate at which undetected and detected infected
subjects develop life-threatening symptoms, and are likely to be comparable if there is no known specific
treatment that is effective against the disease, otherwise \eqn{\mu} is likely to be larger.
These parameters can be reduced by means of improved therapies and acquisition of immunity against
the virus."

\item \eqn{\tau} "denotes the mortality rate (for infected subjects with life-threatening symptoms) and
can be reduced bymeans of improved therapies and treatments."

\item \eqn{\lambda, \kappa, \xi, \rho} and \eqn{\sigma} "denote the rate of recovery for the five
classes of infected subjects, and may differ significantly if an appropriate treatment for the
disease is known and adopted to diagnosed patients, while are probably comparable otherwise.
These parameters can be increased thanks to improved treatments and acquisition of immunity against the
virus."
}
These descriptions of the model parameters are directly taken from the 2020-03-21 paper "A SIDARTHE model of COVID-19 Epidemic in Italy"
by the COVID19 IRCCS San Matteo Pavia Task Force et al. to mitigate any misinterpretation of the input into the model.

(Note the named list must actually have the greek letter names written out, i.e. \code{alpha}, \code{beta},...)

Finally the list \code{initial_conditions} must contain the following, all as fractions of the total population
\itemize{
\item \code{s0} the initial level of susceptible individuals
\item \code{i0} the initial level of infected individuals
\item \code{d0} the initial level of diagnosed individuals
\item \code{a0} the initial level of ailing individuals
\item \code{r0} the initial level of recognized individuals
\item \code{thr0} the initial level of threatened individuals
\item \code{h0} the initial level of heal individuals
\item \code{e0} the initial level of extinct individuals}

It is best to specify all but \code{s0} and then set the initial susceptible fraction to 1 less the sum of the rest.
}
}
\references{
{The Euler scheme implemented by this function is for the model developed in the following paper. It is 8 coupled non-linear ODEs describing dynamics of sub-populations under a pandemic.
The descriptions within the novel paper were used to document the initial conditions and parameters of this model in this package.
\itemize{\item "A SIDARTHE Model of COVID-19 Epidemic in Italy" by the COVID19 IRCCS San Matteo Pavia Task Force et al. \href{https://arxiv.org/abs/2003.09861}{https://arxiv.org/abs/2003.09861}}}
}
