% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir}
\alias{sir}
\title{Euler scheme for the SIR model}
\usage{
sir(
  beta,
  gamma,
  s0 = 1,
  i0 = 1e-04,
  r0 = 0,
  t0 = 0,
  tn = 1,
  n = 1000,
  verbose = TRUE
)
}
\arguments{
\item{beta}{the mean rate of contact, 1/beta the mean time between contact for a given individual (homogeneous population and mixing)}

\item{gamma}{the mean rate of recovery, 1/gamma the mean time until recovery.}

\item{s0}{the initial susceptible population}

\item{i0}{the initial infected population}

\item{r0}{the initial recovering population, defaults to zer0}

\item{t0}{initial time point}

\item{tn}{the terminal time point}

\item{n}{number of sub-intervals in time-grid}

\item{verbose}{whether to plot the populations over time and print parameter+reproductive ratio}
}
\value{
list
\itemize{
\item parameters, a data.frame containing the contact rate, recovery rate, and r0 the basic reproduction ratio.
\item sir, a matrix of the three populations in columns over time, susceptible, infected, recovered}
}
\description{
{A basic Euler scheme is used to numerically solve the ODE system of the standard SIR model.}
}
\examples{
{sir.dat <- sir(beta = 1/2, gamma = 1/3, s0 = 50, i0 = 1, tn = 90, n = 1000)}
}
